Added endpoints for SurpriseBagController and OrderController

+++ src/main/java/dailyfarm/business/entity/LocationValue.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data @Embeddable
 public class LocationValue {
 
-    long latitude;
-    long longitude;
+    double latitude;
+    double longitude;
 }
